requirements.txt
streamlit==1.32.0
pandas==2.1.0
matplotlib==3.8.0
numpy==1.26.0
PyMuPDF==1.23.7
python-docx==1.0.1
docx2txt==0.8
spacy==3.7.2
nltk==3.8.1
flask==2.3.3
werkzeug==2.3.7
scikit-learn==1.3.0
requests==2.31.0
opencv-python==4.8.0.76
pillow==10.0.0